pyproject.toml
[tool.poetry]
name = "fastapi-app-test"
version = "0.1.0"
description = ""
authors = ["vochidat0 <dat.usth2000@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
kubernetes = "^29.0.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
